# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Home automation backend API for IoT state management, alerting, and internet monitoring.",
        "title": "hmauto API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/states": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all states across every type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "states"
                ],
                "summary": "List all states",
                "responses": {
                    "200": {
                        "description": "List of states",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.JsonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/hmstt.StateResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    }
                }
            }
        },
        "/states/{type}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all states for a given type (e.g. switch)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "states"
                ],
                "summary": "List states by type",
                "parameters": [
                    {
                        "type": "string",
                        "example": "switch",
                        "description": "State type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of states",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.JsonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/hmstt.StateResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    },
                    "404": {
                        "description": "No states found for type",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    }
                }
            }
        },
        "/states/{type}/{key}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the state for a specific type and key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "states"
                ],
                "summary": "Get a single state",
                "parameters": [
                    {
                        "type": "string",
                        "example": "switch",
                        "description": "State type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "modem",
                        "description": "State key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "State entry",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.JsonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hmstt.StateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    },
                    "404": {
                        "description": "State not found",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates or updates the state for a specific type and key. Valid types: switch (values: on, off)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "states"
                ],
                "summary": "Set a state value",
                "parameters": [
                    {
                        "type": "string",
                        "example": "switch",
                        "description": "State type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "modem",
                        "description": "State key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "State value",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hmstt.SetStateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated state",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.JsonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hmstt.StateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid type/key/value",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "hmstt.SetStateRequest": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "string",
                    "example": "on"
                }
            }
        },
        "hmstt.StateResponse": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string",
                    "example": "modem"
                },
                "type": {
                    "type": "string",
                    "example": "switch"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-03-16T12:34:56Z"
                },
                "value": {
                    "type": "string",
                    "example": "on"
                }
            }
        },
        "response.JsonResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminKeyAuth": {
            "description": "Format: Bearer {admin-key}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyAuth": {
            "description": "Format: Bearer {api-key}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}